Don't hook SetStdHandle.
It was wrong idea to hook it. We interested only to output to handle that
was active at the process startup (actually to all handles that point to
this output object). Setting StdHandle to handle to any other kernel object
will result just in redirecting std output to that object, but will not affect
stdout that observed by process creator.
That fixes problem with lot output of gperf.exe during re-producing cache hit
on Windows Server 2012 (for some readon msvcrt.dll here calls SetStdHandle
when gperf.exe calls freopen on it output file).